Fix bounds checks and extend fft2 for >2D arrays
Cast _ndim to signed type in Array bounds checks to avoid signed/unsigned comparison warnings/errors. Refactor FFTW::fft2: keep the optimized 2D path (using FFTW plan_dft_2d, shifts, normalization, and plan mutex protection) and add support for input arrays with more than two dimensions by performing a 2D FFT over the last two dimensions for each slice. Also add an input dimensionality check (must be at least 2) and brief comment about multi-dimensional support.

+++ gpi/include/GPIArray/Array.hpp
@@ -741,7 +741,7 @@ public:
 
         #ifdef GPIARRAY_ENABLE_BOUNDS_CHECKS
             // Check if the effective index is within valid bounds (0 to _ndim - 1)
-            if (effective_dim_idx < 0 || effective_dim_idx >= _ndim) {
+            if (effective_dim_idx < 0 || effective_dim_idx >= static_cast<long long>(_ndim)) {
                 THROW_INDEX_ERROR("Axis index " + std::to_string(i) + " (effective " + std::to_string(effective_dim_idx) + ") is out of range for array with " + std::to_string(_ndim) + " dimensions.");
             }
         #endif
@@ -759,7 +759,7 @@ public:
         }
 
         #ifdef GPIARRAY_ENABLE_BOUNDS_CHECKS
-            if (effective_dim_idx < 0 || effective_dim_idx >= _ndim) {
+            if (effective_dim_idx < 0 || effective_dim_idx >= static_cast<long long>(_ndim)) {
                 THROW_INDEX_ERROR("Dimension index " + std::to_string(i) + " (effective " + std::to_string(effective_dim_idx) + ") is out of range for array with " + std::to_string(_ndim) + " dimensions.");
             }
         #endif
